Remove any js/ts duplicates as the result of copying from dummy app
When the dummy app has e.g. a `app.js` file, and the app blueprint (with `--typescript`) has an `app.ts`, after copying the files from the dummy app you would end up with duplicates of `app.js` vs. `app.ts`, which is bad. As the .js file coming from the dummy app could contain meaningful code that is not included in the fresh boilerplate, it should take precedence.